# services/api-gateway/docs/swagger.yaml
basePath: /
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost
info:
  contact: {}
  description: API для платежной системы
  title: Payment API
  version: "1.0"
paths:
...
  /user/all:
    get:
      description: Возвращает список всех зарегистрированных пользователей
      parameters:
      - description: Лимит (по умолчанию нет лимита)
        in: query
        name: limit
        type: integer
      - description: Смещение (по умолчанию 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Получить всех пользователей
      tags:
      - user
  /user/profile:
    get:
      description: Возвращает профиль текущего аутентифицированного пользователя
      parameters:
      - description: ID пользователя
        in: query
        name: userID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Получить профиль
      tags:
      - user
  /user/profile/changeemail:
    patch:
      consumes:
      - application/x-www-form-urlencoded
      description: Изменяет Email текущего аутентифицированного пользователя
      parameters:
      - description: ID пользователя
        in: query
        name: userID
        type: string
      - description: User email
        in: formData
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Изменить Email
      tags:
      - user
  /user/profile/changename:
    patch:
      consumes:
      - application/x-www-form-urlencoded
      description: Изменяет имя текущего аутентифицированного пользователя
      parameters:
      - description: ID пользователя
        in: query
        name: userID
        type: string
      - description: User name
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Изменить имя
      tags:
      - user
  /user/profile/changepassword:
    patch:
      consumes:
      - application/x-www-form-urlencoded
      description: Обновляет пароль текущего аутентифицированного пользователя
      parameters:
      - description: User password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Изменить пароль
      tags:
      - user
schemes:
- https
securityDefinitions:
  CookieAuth:
    description: Аутентификация через куки (токен автоматически отправляется сервером)
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"